Refactored character leveler. refactored DnD5eCharacter leveler. added DnD5eCharacter leveler dialog serise. added character level up dialog window and vm
