+++ PCCharacterManager/DialogWindows/DnD5eLevelupCharacterDialogWindow.xaml
@@ -1,0 +1,18 @@
+﻿<Window
+    x:Class="PCCharacterManager.DialogWindows.DnD5eLevelupCharacterDialogWindow"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="clr-namespace:PCCharacterManager.DialogWindows"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    Title="DnD5eLevelupCharacterDialogWindow"
+    Width="800"
+    Height="450"
+    Background="{DynamicResource 1}"
+    mc:Ignorable="d">
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition />
+        </Grid.ColumnDefinitions>
+    </Grid>
+</Window>
