+++ PCCharacterManager/Models/StarfinderCharacterLeveler.cs
@@ -10,31 +10,14 @@ namespace PCCharacterManager.Models
 {
 	class StarfinderCharacterLeveler : CharacterLeveler
 	{
-		public StarfinderCharacterLeveler(DialogService dialogService) : base(dialogService)
+		public StarfinderCharacterLeveler(DialogServiceBase dialogService) : base(dialogService)
 		{
 		}
 
-		protected override AddClassHelper AddClass(DnD5eCharacter character, string classToAddName)
+		public override void LevelCharacter(DnD5eCharacter character)
 		{
-			throw new NotImplementedException();
-		}
-
-		protected override bool MeetsPrerequisites(DnD5eCharacter character, CharacterMultiClassData characterMultiClassData)
-		{
-			throw new NotImplementedException();
-		}
-
-		protected override void UnLockClassFeatures(DnD5eCharacter character, string className, int classLevel)
-		{
-			throw new NotImplementedException();
-		}
-
-		protected override MultiClass UpdateMaxHealth(DnD5eCharacter character)
-		{
-			MessageBox.Show("Automatic Starfinder character leveling is not yet implemented.", 
+			MessageBox.Show("Automatic Starfinder character leveling is not yet implemented.",
 				"Action Not Yet Supported.", MessageBoxButton.OK, MessageBoxImage.Information);
-
-			return new MultiClass();
 		}
 	}
 }
