+++ PCCharacterManager/ViewModels/Character/CharacterInventoryViewModel.cs
@@ -118,7 +118,7 @@ namespace PCCharacterManager.ViewModels
 
 		public ICommand ShowPropertiesToDisplayCommand { get; }
 
-		public CharacterInventoryViewModel(CharacterStore characterStore, DialogService dialogService)
+		public CharacterInventoryViewModel(CharacterStore characterStore, DialogServiceBase dialogService)
 		{
 			characterStore.SelectedCharacterChange += OnCharacterChanged;
 			Inventory = characterStore.SelectedCharacter.Inventory;
@@ -162,7 +162,7 @@ namespace PCCharacterManager.ViewModels
 				nameof(ItemViewModel.DisplayItemCategory));
 		}
 
-		public CharacterInventoryViewModel(ObservableCollection<ItemViewModel> itemsToDisplay, DialogService dialogService)
+		public CharacterInventoryViewModel(ObservableCollection<ItemViewModel> itemsToDisplay, DialogServiceBase dialogService)
 		{
 			_propertyVMPool = new PropertyEditableVMPool(5);
 			_itemSearch = new ItemSearch();
