+++ PCCharacterManager/ViewModels/TabControlViewModel.cs
@@ -17,7 +17,7 @@ namespace PCCharacterManager.ViewModels
 		public CharacterSpellBookViewModel SpellBookVM { get; private set; }
 		public CharacterNoteBookViewModel NotesVM { get; private set; }
 
-		public TabControlViewModel(CharacterStore characterStore, ICharacterDataService dataService, DialogService dialogService)
+		public TabControlViewModel(CharacterStore characterStore, ICharacterDataService dataService, DialogServiceBase dialogService)
 		{
 			CharacterListVM = new CharacterListViewModel(characterStore, dataService, dialogService);
 			CharacterStatsVM = new CharacterStatsViewModel(characterStore, dialogService);
